backend/requirements.txt
fastapi==0.95.2
uvicorn[standard]==0.29.0
sqlalchemy==2.0.29
pymysql==1.1.0
alembic==1.13.1
python-dotenv==1.0.1
pydantic==1.10.13
